pyproject.toml
[project]
name = "whisperx-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "google-genai>=1.70.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "supabase>=2.28.3",
    "uvicorn>=0.43.0",
    "whisperx>=3.8.4",
]
